backend/requirements.txt
flask
flask-cors
flask-sqlalchemy
requests
PyJWT
werkzeug
reportlab
python-dotenv
